move interface to tests
