+++ src/test/java/com/invisiblecollector/model/builder/IThrowingBuilder.java
@@ -1,4 +1,4 @@
-package com.invisiblecollector.connection.builders;
+package com.invisiblecollector.model.builder;
 
 import com.invisiblecollector.exceptions.IcException;
